+++ internal/defaultervalidator/defaultervalidator.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package webhook
+package defaultervalidator
 
 import (
 	"context"
@@ -42,15 +42,15 @@ var (
 // +kubebuilder:webhook:path=/webhooks/mutating,mutating=true,failurePolicy=fail,sideEffects=None,groups=authentication.gardener.cloud,resources=openidconnects,verbs=create;update,versions=v1alpha1,name=oidc.authentication.gardener.cloud,admissionReviewVersions={v1,v1beta1}
 // +kubebuilder:webhook:verbs=create;update,path=/webhooks/validating,mutating=false,failurePolicy=fail,sideEffects=None,groups=authentication.gardener.cloud,resources=openidconnects,versions=v1alpha1,name=oidc.authentication.gardener.cloud,admissionReviewVersions={v1,v1beta1}
 
-// Handler implements defaulting and validation for the OpenIDConnect resource.
-type Handler struct{}
+// DefaulterValidator implements defaulting and validation for the OpenIDConnect resource.
+type DefaulterValidator struct{}
 
 // ensure webhookHandler implements CustomDefaulter and CustomValidator interfaces
-var _ webhook.CustomDefaulter = (*Handler)(nil)
-var _ webhook.CustomValidator = (*Handler)(nil)
+var _ webhook.CustomDefaulter = (*DefaulterValidator)(nil)
+var _ webhook.CustomValidator = (*DefaulterValidator)(nil)
 
 // Default implements [webhook.CustomDefaulter] so a webhook can be registered for the type.
-func (*Handler) Default(_ context.Context, obj runtime.Object) error {
+func (*DefaulterValidator) Default(_ context.Context, obj runtime.Object) error {
 	oidc, ok := obj.(*authenticationv1alpha1.OpenIDConnect)
 	if !ok {
 		return fmt.Errorf("expected *authenticationv1alpha1.OpenIDConnect but got %T", obj)
@@ -81,7 +81,7 @@ func (*Handler) Default(_ context.Context, obj runtime.Object) error {
 // ValidateCreate validates the object on creation.
 // Return an error if the object is invalid.
 // ValidateCreate implements [webhook.CustomValidator] so a webhook can be registered for the type.
-func (*Handler) ValidateCreate(_ context.Context, obj runtime.Object) (admission.Warnings, error) {
+func (*DefaulterValidator) ValidateCreate(_ context.Context, obj runtime.Object) (admission.Warnings, error) {
 	oidc, ok := obj.(*authenticationv1alpha1.OpenIDConnect)
 	if !ok {
 		return nil, fmt.Errorf("expected *authenticationv1alpha1.OpenIDConnect but got %T", obj)
@@ -95,7 +95,7 @@ func (*Handler) ValidateCreate(_ context.Context, obj runtime.Object) (admission
 // ValidateUpdate validates the object on update.
 // Return an error if the object is invalid.
 // ValidateUpdate implements [webhook.CustomValidator] so a webhook can be registered for the type.
-func (*Handler) ValidateUpdate(_ context.Context, _, obj runtime.Object) (admission.Warnings, error) {
+func (*DefaulterValidator) ValidateUpdate(_ context.Context, _, obj runtime.Object) (admission.Warnings, error) {
 	oidc, ok := obj.(*authenticationv1alpha1.OpenIDConnect)
 	if !ok {
 		return nil, fmt.Errorf("expected *authenticationv1alpha1.OpenIDConnect but got %T", obj)
@@ -109,7 +109,7 @@ func (*Handler) ValidateUpdate(_ context.Context, _, obj runtime.Object) (admiss
 // ValidateDelete validates the object on deletion.
 // Return an error if the object is invalid.
 // ValidateDelete implements [webhook.CustomValidator] so a webhook can be registered for the type.
-func (*Handler) ValidateDelete(_ context.Context, obj runtime.Object) (admission.Warnings, error) {
+func (*DefaulterValidator) ValidateDelete(_ context.Context, obj runtime.Object) (admission.Warnings, error) {
 	oidc, ok := obj.(*authenticationv1alpha1.OpenIDConnect)
 	if !ok {
 		return nil, fmt.Errorf("expected *authenticationv1alpha1.OpenIDConnect but got %T", obj)
