+++ internal/defaultervalidator/defaultervalidator_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package webhook_test
+package defaultervalidator_test
 
 import (
 	"context"
@@ -14,55 +14,55 @@ import (
 	"k8s.io/utils/ptr"
 
 	authenticationv1alpha1 "github.com/gardener/oidc-webhook-authenticator/apis/authentication/v1alpha1"
-	"github.com/gardener/oidc-webhook-authenticator/internal/webhook"
+	"github.com/gardener/oidc-webhook-authenticator/internal/defaultervalidator"
 )
 
 var _ = Describe("OpenidconnectWebhook", func() {
 	var (
-		oidc         *authenticationv1alpha1.OpenIDConnect
-		systemPrefix = authenticationv1alpha1.SystemPrefix
-		handler      webhook.Handler
-		ctx          context.Context
+		oidc               *authenticationv1alpha1.OpenIDConnect
+		systemPrefix       = authenticationv1alpha1.SystemPrefix
+		defaulterValidator defaultervalidator.DefaulterValidator
+		ctx                context.Context
 	)
 
 	BeforeEach(func() {
 		oidc = &authenticationv1alpha1.OpenIDConnect{}
-		handler = webhook.Handler{}
+		defaulterValidator = defaultervalidator.DefaulterValidator{}
 		ctx = context.Background()
 	})
 
 	Context("defaulting", func() {
 		It("should default username claim", func() {
-			Expect(handler.Default(ctx, oidc)).To(Succeed())
+			Expect(defaulterValidator.Default(ctx, oidc)).To(Succeed())
 			Expect(*oidc.Spec.UsernameClaim).To(Equal("sub"))
 		})
 
 		It("should not default username claim if explicitly set", func() {
 			oidc.Spec.UsernameClaim = ptr.To("someuserclaim")
-			Expect(handler.Default(ctx, oidc)).To(Succeed())
+			Expect(defaulterValidator.Default(ctx, oidc)).To(Succeed())
 			Expect(*oidc.Spec.UsernameClaim).To(Equal("someuserclaim"))
 		})
 
 		It("should default groups claim", func() {
-			Expect(handler.Default(ctx, oidc)).To(Succeed())
+			Expect(defaulterValidator.Default(ctx, oidc)).To(Succeed())
 			Expect(*oidc.Spec.GroupsClaim).To(Equal("groups"))
 		})
 
 		It("should not default groups claim if explicitly set", func() {
 			oidc.Spec.GroupsClaim = ptr.To("somegroupsclaim")
-			Expect(handler.Default(ctx, oidc)).To(Succeed())
+			Expect(defaulterValidator.Default(ctx, oidc)).To(Succeed())
 			Expect(*oidc.Spec.GroupsClaim).To(Equal("somegroupsclaim"))
 		})
 
 		It("should default supported signing algs", func() {
-			Expect(handler.Default(ctx, oidc)).To(Succeed())
+			Expect(defaulterValidator.Default(ctx, oidc)).To(Succeed())
 			Expect(len(oidc.Spec.SupportedSigningAlgs)).To(Equal(1))
 			Expect(oidc.Spec.SupportedSigningAlgs[0]).To(Equal(authenticationv1alpha1.RS256))
 		})
 
 		It("should not default supported signing algs if explicitly set", func() {
 			oidc.Spec.SupportedSigningAlgs = []authenticationv1alpha1.SigningAlgorithm{authenticationv1alpha1.RS256, authenticationv1alpha1.RS512}
-			Expect(handler.Default(ctx, oidc)).To(Succeed())
+			Expect(defaulterValidator.Default(ctx, oidc)).To(Succeed())
 			Expect(len(oidc.Spec.SupportedSigningAlgs)).To(Equal(2))
 			Expect(oidc.Spec.SupportedSigningAlgs).To(ConsistOf(authenticationv1alpha1.RS256, authenticationv1alpha1.RS512))
 		})
@@ -76,7 +76,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 
 		It("should return error if issuer url is not starting with https", func() {
 			oidc.Spec.IssuerURL = "http://notsecure.com"
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("issuerURL: Invalid value: \"http://notsecure.com\": must start with https"))
 			Expect(warnings).To(BeNil())
@@ -88,7 +88,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 				"deprecated1",
 				"deprecated2",
 			}
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("[supportedSigningAlgs[1]: Invalid value: \"deprecated1\": unsupported signing alg, supportedSigningAlgs[2]: Invalid value: \"deprecated2\": unsupported signing alg]"))
 			Expect(warnings).To(BeNil())
@@ -96,7 +96,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 
 		It("should return error for invalid ca bundle", func() {
 			oidc.Spec.CABundle = []byte("dGVzdA==")
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("caBundle: Invalid value: \"dGVzdA==\": must be a valid base64 encoded certificate bundle"))
 			Expect(warnings).To(BeNil())
@@ -104,7 +104,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 
 		It("should return error for invalid jwks", func() {
 			oidc.Spec.JWKS.Keys = []byte("dGVzdA==")
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("keys: Invalid value: \"dGVzdA==\": must be a valid base64 encoded JWKS"))
 			Expect(warnings).To(BeNil())
@@ -112,7 +112,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 
 		It("should return error for empty clientID", func() {
 			oidc.Spec.ClientID = ""
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("clientID: Invalid value: \"\": must not be empty"))
 			Expect(warnings).To(BeNil())
@@ -122,7 +122,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 			func(maliciousPrefix string) {
 				errMessageTemplate := "usernamePrefix: Invalid value: \"%s\": must not start with system:"
 				oidc.Spec.UsernamePrefix = &maliciousPrefix
-				warnings, err := handler.ValidateCreate(ctx, oidc)
+				warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 				Expect(err).To(HaveOccurred())
 				Expect(err.Error()).To(Equal(fmt.Sprintf(errMessageTemplate, maliciousPrefix)))
 				Expect(warnings).To(BeNil())
@@ -135,7 +135,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 		DescribeTable("should allow username prefix if it does not start with 'system:'",
 			func(validPrefix string) {
 				oidc.Spec.UsernamePrefix = &validPrefix
-				warnings, err := handler.ValidateCreate(ctx, oidc)
+				warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(warnings).To(BeNil())
 			},
@@ -151,7 +151,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 			func(maliciousPrefix string) {
 				errMessageTemplate := "groupsPrefix: Invalid value: \"%s\": must not start with system:"
 				oidc.Spec.GroupsPrefix = &maliciousPrefix
-				warnings, err := handler.ValidateCreate(ctx, oidc)
+				warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 				Expect(err).To(HaveOccurred())
 				Expect(err.Error()).To(Equal(fmt.Sprintf(errMessageTemplate, maliciousPrefix)))
 				Expect(warnings).To(BeNil())
@@ -164,7 +164,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 		DescribeTable("should allow groups prefix if it does not start with 'system:'",
 			func(validPrefix string) {
 				oidc.Spec.GroupsPrefix = &validPrefix
-				warnings, err := handler.ValidateCreate(ctx, oidc)
+				warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(warnings).To(BeNil())
 			},
@@ -190,7 +190,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 					ClientID:  "some-id",
 				},
 			}
-			warnings, err := handler.ValidateUpdate(ctx, oidc, &newObj)
+			warnings, err := defaulterValidator.ValidateUpdate(ctx, oidc, &newObj)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(warnings).To(BeNil())
 		})
@@ -202,7 +202,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 					ClientID:  "some-id",
 				},
 			}
-			warnings, err := handler.ValidateUpdate(ctx, oidc, &newObj)
+			warnings, err := defaulterValidator.ValidateUpdate(ctx, oidc, &newObj)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("issuerURL: Invalid value: \"http://notsecure.com\": must start with https"))
 			Expect(warnings).To(BeNil())
@@ -218,7 +218,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 						UsernamePrefix: &maliciousPrefix,
 					},
 				}
-				warnings, err := handler.ValidateUpdate(ctx, oidc, &newObj)
+				warnings, err := defaulterValidator.ValidateUpdate(ctx, oidc, &newObj)
 				Expect(err).To(HaveOccurred())
 				Expect(err.Error()).To(Equal(fmt.Sprintf(errMessageTemplate, maliciousPrefix)))
 				Expect(warnings).To(BeNil())
@@ -237,7 +237,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 						UsernamePrefix: &validPrefix,
 					},
 				}
-				warnings, err := handler.ValidateUpdate(ctx, oidc, &newObj)
+				warnings, err := defaulterValidator.ValidateUpdate(ctx, oidc, &newObj)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(warnings).To(BeNil())
 			},
@@ -259,7 +259,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 						GroupsPrefix: &maliciousPrefix,
 					},
 				}
-				warnings, err := handler.ValidateUpdate(ctx, oidc, &newObj)
+				warnings, err := defaulterValidator.ValidateUpdate(ctx, oidc, &newObj)
 				Expect(err).To(HaveOccurred())
 				Expect(err.Error()).To(Equal(fmt.Sprintf(errMessageTemplate, maliciousPrefix)))
 				Expect(warnings).To(BeNil())
@@ -278,7 +278,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 						GroupsPrefix: &validPrefix,
 					},
 				}
-				warnings, err := handler.ValidateUpdate(ctx, oidc, &newObj)
+				warnings, err := defaulterValidator.ValidateUpdate(ctx, oidc, &newObj)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(warnings).To(BeNil())
 			},
@@ -292,7 +292,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 
 		It("should return error for negative maxTokenExpirationSeconds", func() {
 			oidc.Spec.MaxTokenExpirationSeconds = ptr.To[int64](-1)
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("maxTokenExpirationSeconds: Invalid value: -1: should be positive"))
 			Expect(warnings).To(BeNil())
@@ -300,7 +300,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 
 		It("should return error for zero maxTokenExpirationSeconds", func() {
 			oidc.Spec.MaxTokenExpirationSeconds = ptr.To[int64](0)
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("maxTokenExpirationSeconds: Invalid value: 0: should be positive"))
 			Expect(warnings).To(BeNil())
@@ -308,7 +308,7 @@ var _ = Describe("OpenidconnectWebhook", func() {
 
 		It("should return error for duplicate claims in ExtraClaims", func() {
 			oidc.Spec.ExtraClaims = []string{"claim1", "claim2", "ClaIm1"}
-			warnings, err := handler.ValidateCreate(ctx, oidc)
+			warnings, err := defaulterValidator.ValidateCreate(ctx, oidc)
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("extraClaims: Invalid value: []string{\"claim1\", \"claim2\", \"ClaIm1\"}: duplicated claims found"))
 			Expect(warnings).To(BeNil())
